clients/pred_pc_grpc_py/requirements.txt
grpcio==1.46.3
tensorflow==2.9.1
tensorflow-cpu==2.9.1
tensorflow-serving-api==2.9.1
